move controller logic to controllers folder
