--- src/pepr/operator/network/README.md
@@ -1,1 +1,0 @@
-This folder contains UDS Core common network policies for the cluster. They establish a baseline security posture that should be used in conjunction with individual allow policies for each workload.
